Fixed an issue where the cube shadow maps were being recreated on window resize. These have static size. Therefore, no need to recreate them.
The above fixed an issue where resizing the window would result in the application terminating. This is because the descriptor set that references the cube shadow map wasn't being updated to reference the recreated cube shadow map.

Fixed an issue where we were using uin32_t to subscript an array in our torus and sphere generation.

create_sampler_info was default initializing the comparison operation to VK_COMPARE_OP_NEVER. This has been changed to VK_COMPARE_OP_GREATER_OR_EQUAL (our depth is reversed).

 Added additional PCF offsets.

+++ kleicha/Utils.h
@@ -32,8 +32,8 @@ namespace utils {
 
 //    vkt::IndexedMesh generate_cube_mesh();
     vkt::Mesh generate_pyramid_mesh();
-    vkt::Mesh generate_sphere(uint32_t prec);
-    vkt::Mesh generate_torus(uint32_t prec, float inner, float outer);
+    vkt::Mesh generate_sphere(size_t prec);
+    vkt::Mesh generate_torus(size_t prec, float inner, float outer);
     vkt::Mesh load_obj_mesh(const char* filePath, vkt::MeshType meshType);
 
     vkt::Buffer create_buffer(VmaAllocator allocator, VkDeviceSize bufferSize, VkBufferUsageFlags bufferUsage,
